Implementation control exception with @ExceptionHandler at class GlobalExceptionHandler, MethodArgumentNotValidException OK and ResourceNotFoundException OK at PetController, ImageController and AuthController. Creation endpoint Get user profile.

+++ BackEnd/src/main/java/com/mascotas/back/controller/ImageController.java
@@ -1,12 +1,15 @@
 package com.mascotas.back.controller;
 
 import com.mascotas.back.dto.ImageDto;
+import com.mascotas.back.dto.ImageResponseDto;
+import com.mascotas.back.exception.ResourceNotFoundException;
 import com.mascotas.back.model.Image;
 import com.mascotas.back.service.ImageService;
 
 import java.net.URI;
 import java.util.List;
 
+import com.mascotas.back.service.PetService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -30,50 +33,61 @@ import org.springframework.web.util.UriComponentsBuilder;
 public class ImageController {
 
     private final ImageService imageService;
+    private final PetService petService;
 
     @DeleteMapping("/image/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteImage(@PathVariable Long id) {
+        if (!imageService.existsById(id)) throw new ResourceNotFoundException("image", "id", id);
         imageService.deleteImage(id);
     }
 
     @GetMapping("/images/{pet_id}")
-    public ResponseEntity<List<ImageDto>> listImages(@PathVariable Long pet_id) {
+    public ResponseEntity<List<ImageResponseDto>> listImages(@PathVariable Long pet_id) {
         List<Image> images = imageService.listImage(pet_id);
-        List<ImageDto> imageDtoList = images.stream().map( // Iterando sobre la lista de tipo Image para obtener el campo "byte[] image"
-                                                        image -> new ImageDto(image.getImage())
-                                                        ).toList();
+        if (!petService.existsById(pet_id)) throw new ResourceNotFoundException("pet", "pet_id", pet_id);
+        if (images == null || images.isEmpty()) throw new ResourceNotFoundException("images");
+        List<ImageResponseDto> imageDtoList = images.stream().map(  // Convertir una lista de tipo Image a lista de tipo ImageResponseDto con patrón Builder
+                                                                    image -> ImageResponseDto
+                                                                            .builder()
+                                                                            .image_id(image.getId())
+                                                                            .imageBase64(image.getImage())
+                                                                            .build()
+                                                                  ).toList();
         return ResponseEntity.ok(imageDtoList);
     }
 
     @GetMapping("/image/{image_id}")
     public ResponseEntity<ImageDto> getImage(@PathVariable Long image_id) {
         ImageDto imageDto = imageService.findImageById(image_id);
+        if (imageDto == null) throw new ResourceNotFoundException("image", "image_id", image_id);
         return ResponseEntity.ok(imageDto);
     }
 
     @PostMapping("/image/{pet_id}")
-    public ResponseEntity<ImageDto> createImage(@RequestBody @Valid ImageDto imageDto, @PathVariable Long pet_id, UriComponentsBuilder uriComponentsBuilder){
+    public ResponseEntity<ImageResponseDto> createImage(@RequestBody @Valid ImageDto imageDto, @PathVariable Long pet_id, UriComponentsBuilder uriComponentsBuilder){
+        if (!petService.existsById(pet_id)) throw new ResourceNotFoundException("pet", "pet_id", pet_id);
         Image image = imageService.saveImage(imageDto.imageBase64, pet_id);
         URI url = uriComponentsBuilder.path("api/v1/image/{image_id}").buildAndExpand(image.getId()).toUri();
-        ImageDto imageResponse = ImageDto.builder()
-                .imageBase64(image.getImage())
-                .build();
+        ImageResponseDto imageResponse = ImageResponseDto
+                                            .builder()
+                                            .image_id(image.getId())
+                                            .imageBase64(image.getImage())
+                                            .build();
         return ResponseEntity.created(url).body(imageResponse);
     }
 
     @PutMapping("/image/{image_id}")
-    public  ResponseEntity<ImageDto> updateImage(@RequestBody @Valid ImageDto imageDto, @PathVariable Long image_id) {
+    public ResponseEntity<ImageDto> updateImage(@RequestBody @Valid ImageDto imageDto, @PathVariable Long image_id) {
        if (imageService.existsById(image_id)) {
             Image image = imageService.updateImage(imageDto.imageBase64, image_id);
-            //Ejemplo para los create con Dto
             ImageDto imageResponseDto = ImageDto
                     .builder()
                     .imageBase64(image.getImage())
                     .build();
             return ResponseEntity.ok(imageResponseDto);
         } else {
-            return null; // Crear excepción personalizada NOT_FOUND (id no existe)
+           throw new ResourceNotFoundException("image", "image_id", image_id); // Crear excepción personalizada NOT_FOUND (id no existe)
         }
     }
